Work on big update

+++ crates/nox/src/gpu/image.rs
@@ -1,17 +1,13 @@
 mod format;
-mod enums;
-mod structs;
 mod state;
-mod attributes;
+mod create_info;
 mod properties;
-mod sampler;
 mod error;
 
-use crate::sync::Arc;
-
-use core::ptr::NonNull;
-
-use core::hash::{Hash, Hasher};
+use core::{
+    hash::{Hash, Hasher},
+    ptr::NonNull,
+};
 
 use nox_ash::vk;
 
@@ -29,31 +25,28 @@ use nox_mem::{
 
 use nox_alloc::arena::{self, CellArena};
 
-use crate::gpu::prelude::{memory_binder::{DeviceMemory, MemoryBinder}, *};
+use {
+    crate::gpu::prelude::{
+        memory_binder::{DeviceMemory, MemoryBinder},
+        subresource_state::*,
+        *
+    },
+    crate::dev::{has_bits, has_not_bits},
+    crate::dev::error as dev_error,
+    crate::sync::Arc,
+};
 
-use crate::dev::{has_bits, has_not_bits};
-
-use crate::dev::error as dev_error;
-
-use super::{MSAA, Vulkan};
+use super::{MsaaSamples, Vulkan};
 
 pub use format::*;
-pub use enums::*;
-pub use structs::*;
 pub use error::*;
-pub use attributes::*;
-pub use sampler::*;
+pub use create_info::*;
 pub(crate) use state::*;
 pub(crate) use properties::ImageProperties;
 
-pub struct Image;
-
-impl Image {
-
-    #[inline(always)]
-    pub fn default_attributs() -> ImageAttributes {
-        ImageAttributes::new()
-    }
+enum MemorySource {
+    Owned(Box<dyn DeviceMemory>, usize),
+    Swapchain,
 }
 
 pub(crate) struct ImageMeta {
@@ -63,9 +56,8 @@ pub(crate) struct ImageMeta {
     subviews: AHashMap<ImageRange, vk::ImageView>,
     properties: ImageProperties,
     states: NonNullVec32<'static, NonNullVec32<'static, ImageMipRange>>,
-    memory: Box<dyn DeviceMemory>,
+    memory: MemorySource,
     last_used_frame: u64,
-    arena_size: usize,
 }
 
 unsafe impl Send for ImageMeta {}
@@ -73,86 +65,89 @@ unsafe impl Sync for ImageMeta {}
 
 impl ImageMeta {
 
-    pub(crate) fn new(
+    fn new(
         vk: Arc<Vulkan>,
-        attributes: ImageAttributes,
+        create_info: &ImageCreateInfo<'_>,
         alloc: &mut (impl MemoryBinder + ?Sized),
+        bind_memory_info: &mut vk::BindImageMemoryInfo<'static>,
     ) -> Result<Self, dev_error::Error>
     {
         let mut image_type = vk::ImageType::TYPE_2D;
-        if attributes.dimensions.depth > 1 {
-            if attributes.array_layers != 1 {
+        if create_info.dimensions.depth > 1 {
+            if create_info.array_layers != 1 {
                 return Err(dev_error::Error::just_context(CompactString::new(
                     "image layers must be 1 if depth is greater than 1"
                 )))
             }
             image_type = vk::ImageType::TYPE_3D;
         }
-        if attributes.format == vk::Format::UNDEFINED {
+        if create_info.format == vk::Format::UNDEFINED {
             return Err(dev_error::Error::just_context(CompactString::new(
                 "image format must be defined"
             )))
         }
         let mut flags = Default::default();
-        if attributes.mutable_format {
+        if create_info.mutable_format {
             flags |= vk::ImageCreateFlags::MUTABLE_FORMAT;
         }
-        if attributes.cube_map {
+        if create_info.cube_map {
             flags |= vk::ImageCreateFlags::CUBE_COMPATIBLE;
-            if attributes.dimensions.width != attributes.dimensions.height ||
-                attributes.dimensions.depth != 1
+            if create_info.dimensions.width != create_info.dimensions.height ||
+                create_info.dimensions.depth != 1
             {
                 return Err(dev_error::Error::just_context(format_compact!(
                     "width ({}) and height ({}) of a cube map compatible image must be equal and depth ({}) must be 1",
-                    attributes.dimensions.width, attributes.dimensions.height, attributes.dimensions.depth,
+                    create_info.dimensions.width, create_info.dimensions.height, create_info.dimensions.depth,
                 )))
             }
         }
-        if attributes.dimensions.is_zero() {
+        if create_info.dimensions.is_zero() {
             return Err(dev_error::Error::just_context(format_compact!(
                 "image dimensions {} must not be zero",
-                attributes.dimensions,
+                create_info.dimensions,
             )))
         }
-        if attributes.array_layers == 0 {
+        if create_info.array_layers == 0 {
             return Err(dev_error::Error::just_context(format_compact!(
                 "image layers must be greater than 0",
             )))
         }
-        if attributes.cube_map && attributes.array_layers < 6 {
+        if create_info.cube_map && create_info.array_layers < 6 {
             return Err(dev_error::Error::just_context(format_compact!(
                 "layer count {} of a cube map/array image must be at least 6",
-                attributes.array_layers,
+                create_info.array_layers,
             )))
         }
-        if attributes.mip_levels == 0 {
+        if create_info.mip_levels == 0 {
             return Err(dev_error::Error::just_context(format_compact!(
                 "mip levels must be greater than zero",
             )))
         }
-        let create_info = vk::ImageCreateInfo {
+        let vk_create_info = vk::ImageCreateInfo {
             s_type: vk::StructureType::IMAGE_CREATE_INFO,
             flags,
             image_type,
-            format: attributes.format,
-            extent: attributes.dimensions.into(),
-            mip_levels: attributes.mip_levels,
-            array_layers: attributes.array_layers,
-            samples: attributes.samples.into(),
+            format: create_info.format,
+            extent: create_info.dimensions.into(),
+            mip_levels: create_info.mip_levels,
+            array_layers: create_info.array_layers,
+            samples: create_info.samples.into(),
             tiling: vk::ImageTiling::OPTIMAL,
-            usage: attributes.usage,
+            usage: create_info.usage.into(),
             sharing_mode: vk::SharingMode::EXCLUSIVE,
             initial_layout: vk::ImageLayout::UNDEFINED,
             ..Default::default()
         };
-        let mut mem_requirements = Default::default();
+        let mut dedicated_memory_requirements = vk::MemoryDedicatedRequirements::default();
+        let mut mem_requirements = vk::MemoryRequirements2
+            ::default().push_next(&mut dedicated_memory_requirements);
         unsafe {
             let mut device_mem_requirements = vk::DeviceImageMemoryRequirements {
                 s_type: vk::StructureType::DEVICE_IMAGE_MEMORY_REQUIREMENTS,
-                p_create_info: &create_info,
+                p_create_info: &vk_create_info,
                 ..Default::default()
             };
-            vk.maintenance4_device().get_device_image_memory_requirements(
+            vk.device().get_device_image_memory_requirements(
                 &device_mem_requirements, &mut mem_requirements,
             );
         };
@@ -161,25 +156,25 @@ impl ImageMeta {
             .context("failed to allocate GPU memory for image")?
         };
         let handle = unsafe {
-            vk.device().create_image(&create_info, None)
+            vk.device().create_image(&vk_create_info, None)
             .context("failed to create Vulkan image")?
         };
-        unsafe {
-            vk.device().bind_image_memory(
-                handle, memory.device_memory(),
-                memory.offset()
-            ).context("failed to bind image memory")?
+        *bind_memory_info = vk::BindImageMemoryInfo {
+            image: handle,
+            memory: memory.device_memory(),
+            memory_offset: memory.offset(),
+            ..Default::default()
         };
         let properties = ImageProperties {
-            dimensions: attributes.dimensions,
-            format: attributes.format,
-            aspect_mask: attributes.aspects.as_raw(),
-            usage: attributes.usage,
-            samples: attributes.samples,
-            array_layers: attributes.array_layers,
-            mip_levels: attributes.mip_levels,
+            dimensions: create_info.dimensions,
+            format: create_info.format,
+            aspect_mask: create_info.aspects,
+            usage: create_info.usage.into(),
+            samples: create_info.samples,
+            array_layers: create_info.array_layers,
+            mip_levels: create_info.mip_levels,
             create_flags: flags,
-            format_resolve_modes: attributes.resolve_modes,
+            format_resolve_modes: create_info.resolve_modes,
         };
         let view_create_info = vk::ImageViewCreateInfo {
             s_type: vk::StructureType::IMAGE_VIEW_CREATE_INFO,
@@ -212,7 +207,7 @@ impl ImageMeta {
                     access_mask: vk::AccessFlags2::NONE,
                     layout: vk::ImageLayout::UNDEFINED,
                     queue_family_index: vk::QUEUE_FAMILY_IGNORED,
-                    command_index: COMMAND_REQUEST_IGNORED,
+                    command_index: COMMAND_INDEX_IGNORED,
                     command_timeline_value: 0,
                 },
                 base_mip_level: 0,
@@ -227,25 +222,94 @@ impl ImageMeta {
             subviews: AHashMap::default(),
             properties,
             states,
-            memory,
+            memory: MemorySource::Owned(memory, arena.into_inner().size()),
             last_used_frame: 0,
-            arena_size: arena.into_inner().size(),
         })
+    }
+
+    pub unsafe fn from_swapchain_image(
+        vk: Arc<Vulkan>,
+        handle: vk::Image,
+        dimensions: Dimensions,
+        format: vk::Format,
+        usage: vk::ImageUsageFlags,
+        alloc: &impl LocalAlloc<Error = dev_error::Error>
+    ) -> Result<Self, dev_error::Error> {
+        let properties = ImageProperties {
+            dimensions,
+            format,
+            aspect_mask: ImageAspectFlags::COLOR,
+            usage: usage.into(),
+            samples: MsaaSamples::X1,
+            array_layers: 1,
+            mip_levels: 1,
+            create_flags: vk::ImageCreateFlags::empty(),
+            format_resolve_modes: Default::default(),
+        };
+        let view_create_info = vk::ImageViewCreateInfo {
+            s_type: vk::StructureType::IMAGE_VIEW_CREATE_INFO,
+            image: handle,
+            view_type: properties.view_type(), 
+            format: properties.format,
+            components: ComponentMapping::default().into(),
+            subresource_range: properties.whole_subresource().into(),
+            ..Default::default()
+        };
+        let view = unsafe {
+            vk.device().create_image_view(&view_create_info, None)
+            .context("failed to create main image view")?
+        };
+        let mip_levels = properties.mip_levels;
+        let num_layers = properties.aspect_mask.count_ones() * properties.array_layers;
+        let mut states = NonNullVec32
+            ::with_capacity(num_layers, alloc)?
+            .into_static();
+        states.push({
+            let mut vec = NonNullVec32::with_capacity(1, alloc)?;
+            vec.push(ImageMipRange {
+                state: ImageSubresourceState {
+                    stage_mask: vk::PipelineStageFlags2::NONE,
+                    access_mask: vk::AccessFlags2::NONE,
+                    layout: vk::ImageLayout::UNDEFINED,
+                    queue_family_index: vk::QUEUE_FAMILY_IGNORED,
+                    command_index: COMMAND_INDEX_IGNORED,
+                    command_timeline_value: 0,
+                },
+                base_mip_level: 0,
+                level_count: mip_levels,
+            });
+            vec.into_static()
+        });
+        Ok(Self {
+            vk,
+            handle,
+            view,
+            subviews: AHashMap::default(),
+            properties,
+            states,
+            memory: MemorySource::Swapchain,
+            last_used_frame: 0,
+        })
+    }
+
+    #[inline(always)]
+    pub fn is_swapchain(&self) -> bool {
+        matches!(self.memory, MemorySource::Swapchain)
     }
 
     fn get_states_mut(
         &mut self,
-        aspect: vk::ImageAspectFlags,
+        aspect: ImageAspectFlags,
         layer: u32,
     ) -> Option<&mut NonNullVec32<'static, ImageMipRange>>
     {
-        let aspect_mask = self.properties.aspect_mask;
+        let aspect_mask = self.properties.aspect_mask.as_raw();
         let aspect = aspect.as_raw();
-        if aspect_mask & aspect != aspect {
+        if aspect.count_ones() != 1 || aspect_mask & aspect != aspect {
             return None
         }
-        let mut cindex = 0;
-        let mut index = 0;
+        let mut cindex = 0u32;
+        let mut index = 0u32;
         macro_rules! index_op {
             ($($n:expr),* $(,)?) => {
                 $(
@@ -261,6 +325,36 @@ impl ImageMeta {
         index -= 1;
         let index = index * self.properties.array_layers + layer;
         Some(&mut self.states[index as usize])
+    }
+
+    pub fn get_states(
+        &self,
+        aspect: ImageAspectFlags,
+        layer: u32,
+    ) -> Option<&NonNullVec32<'static, ImageMipRange>>
+    {
+        let aspect_mask = self.properties.aspect_mask.as_raw();
+        let aspect = aspect.as_raw();
+        if aspect.count_ones() != 1 || aspect_mask & aspect != aspect {
+            return None
+        }
+        let mut cindex = 0u32;
+        let mut index = 0u32;
+        macro_rules! index_op {
+            ($($n:expr),* $(,)?) => {
+                $(
+                    cindex += ((1 << $n) & aspect_mask) >> $n;
+                    index |= (((1 << $n) & aspect) >> $n).wrapping_neg() & cindex;
+                )*
+            };
+        }
+        index_op!(
+            1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16,
+            /* 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, <- not needed */
+        );
+        index -= 1;
+        let index = index * self.properties.array_layers + layer;
+        Some(&self.states[index as usize])
     }
 
     #[inline(always)]
@@ -284,16 +378,15 @@ impl ImageMeta {
         })
     }
 
-    #[inline(always)]
     pub(crate) fn validate_layers(
         &self,
         layers: ImageSubresourceLayers
     ) -> Option<ImageError>
     {
         if has_not_bits!(self.properties.aspect_mask, layers.aspect_mask) {
-            return Some(ImageError::AspectMismatch(vk::ImageAspectFlags::from_raw(
+            return Some(ImageError::AspectMismatch(
                 layers.aspect_mask ^ self.properties.aspect_mask & layers.aspect_mask
-            )))
+            ))
         }
         if layers.mip_level > self.properties.mip_levels ||
             layers.base_array_layer + layers.layer_count.get() > self.properties.array_layers
@@ -315,7 +408,6 @@ impl ImageMeta {
         self.view
     }
 
-    #[inline(always)]
     pub(crate) fn get_subview(
         &mut self,
         range: ImageRange,
@@ -331,14 +423,15 @@ impl ImageMeta {
                         format: self.properties.format,
                     }
                 };
+            let view_type = self.properties.validate_range(&range)?;
             let device = self.vk.device();
             let create_info = vk::ImageViewCreateInfo {
                 s_type: vk::StructureType::IMAGE_VIEW_CREATE_INFO,
                 image: self.handle,
-                view_type: self.properties.validate_range(range)?,
+                view_type,
                 format: component_info.format,
                 components: component_info.component_mapping.into(),
-                subresource_range: range.subresource.into(),
+                subresource_range: range.subresource_range.into(),
                 ..Default::default()
             };
             let view = unsafe {
@@ -348,7 +441,6 @@ impl ImageMeta {
         })?)
     }
 
-    #[inline(always)]
     pub fn flush_subresources(&mut self) {
         for states in &mut self.states {
             for range in states {
@@ -358,7 +450,6 @@ impl ImageMeta {
         }
     }
 
-    #[inline(always)]
     pub(crate) fn memory_barrier<'a>(
         &mut self,
         state: ImageSubresourceState,
@@ -370,7 +461,7 @@ impl ImageMeta {
         cache.barriers.clear();
         let subresource =
             if let Some(info) = subresource_info {
-                self.properties.validate_range(ImageRange::new(info, None))?;
+                self.properties.validate_range(&ImageRange::new(info, None))?;
                 info
             } else {
                 self.properties.whole_subresource()
@@ -382,9 +473,9 @@ impl ImageMeta {
         };
         let layer_start = subresource.base_array_layer;
         let layer_end = subresource.base_array_layer + subresource.layer_count.get();
-        for aspect in subresource.aspect_mask
+        for aspect in subresource.aspect_mask.as_raw()
             .bit_iter()
-            .map(|aspect| vk::ImageAspectFlags::from_raw(aspect))
+            .map(|aspect| ImageAspectFlags::from_raw(aspect))
         {
             for layer in layer_start..layer_end {
                 let mut not_inserted = None;
@@ -399,7 +490,7 @@ impl ImageMeta {
                         },
                         StateOverwrite::Consume(mut barrier) => {
                             ranges.remove(i);
-                            barrier.subresource_range.aspect_mask = aspect;
+                            barrier.subresource_range.aspect_mask = aspect.into();
                             barrier.subresource_range.base_array_layer = layer;
                             if !preserve_contents {
                                 barrier.old_layout = vk::ImageLayout::UNDEFINED;
@@ -409,7 +500,7 @@ impl ImageMeta {
                         },
                         StateOverwrite::Cut(left, right, mut barrier) => {
                             ranges.remove(i);
-                            barrier.subresource_range.aspect_mask = aspect;
+                            barrier.subresource_range.aspect_mask = aspect.into();
                             barrier.subresource_range.base_array_layer = layer;
                             if !preserve_contents {
                                 barrier.old_layout = vk::ImageLayout::UNDEFINED;
@@ -427,7 +518,7 @@ impl ImageMeta {
                         },
                         StateOverwrite::Shrink(new_range, mut barrier) => {
                             ranges[i as usize] = new_range;
-                            barrier.subresource_range.aspect_mask = aspect;
+                            barrier.subresource_range.aspect_mask = aspect.into();
                             barrier.subresource_range.base_array_layer = layer;
                             if !preserve_contents {
                                 barrier.old_layout = vk::ImageLayout::UNDEFINED;
@@ -495,14 +586,16 @@ impl Drop for ImageMeta {
                 device.destroy_image_view(subview, None);
             }
             device.destroy_image_view(self.view, None);
-            device.destroy_image(self.handle(), None);
-            StdAlloc.free_raw(
-                NonNull::new_unchecked(self.states.as_mut_ptr()).cast(),
-                Layout::from_size_align_unchecked(
-                    self.arena_size,
-                    arena::max_align()
-                ),
-            );
+            if let MemorySource::Owned(_, arena_size) = &self.memory {
+                device.destroy_image(self.handle(), None);
+                StdAlloc.free_raw(
+                    NonNull::new_unchecked(self.states.as_mut_ptr()).cast(),
+                    Layout::from_size_align_unchecked(
+                        *arena_size,
+                        arena::max_align()
+                    ),
+                );
+            }
         }
     }
 }
